Move GetAllPids from OrbitService/Utils to OrbitLinuxTracing/Utils
This is currently unused. Will be used in the upcoming commits.
