+++ OrbitLinuxTracing/Utils.h
@@ -25,6 +25,8 @@ std::optional<std::string> ReadFile(std::string_view filename);
 
 std::string ReadMaps(pid_t pid);
 
+std::vector<pid_t> GetAllPids();
+
 std::vector<pid_t> ListThreads(pid_t pid);
 
 std::string GetThreadName(pid_t tid);
