+++ OrbitService/Utils.h
@@ -24,7 +24,6 @@ ErrorMessageOr<std::vector<orbit_grpc_protos::ModuleInfo>> ReadModules(int32_t p
 ErrorMessageOr<std::vector<orbit_grpc_protos::TracepointInfo>> ReadTracepoints();
 ErrorMessageOr<std::vector<orbit_grpc_protos::ModuleInfo>> ParseMaps(
     std::string_view proc_maps_data);
-std::vector<pid_t> GetAllPids();
 
 // In the linux world, Jiffies is a global counter which increments on tick (caused by a CPU timer
 // interrupt). This struct is a poor man's strong type to ensure that this measure is not mistakenly
